+++ lib/notifications_page.dart
@@ -4,13 +4,33 @@ import 'package:get_time_ago/get_time_ago.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:path/path.dart' as path;
 
+class ThemeConfig {
+  final Color? darkText;
+  final Color? lightText;
+  final Color? darkBackground;
+  final Color? lightBackground;
+  final Color? primary;
+
+  ThemeConfig({
+    this.lightText,
+    this.darkText,
+    this.lightBackground,
+    this.darkBackground,
+    this.primary,
+  });
+}
+
 class FynoNotificationIcon extends StatefulWidget {
   final FynoInApp fynoInApp;
   final Color iconColor;
+  final IconData? notificationIcon;
+  final ThemeConfig? themeConfig;
 
   const FynoNotificationIcon(
     this.fynoInApp,
     this.iconColor, {
+    this.notificationIcon,
+    this.themeConfig,
     Key? key,
   }) : super(key: key);
 
@@ -22,29 +42,33 @@ class FynoNotificationIconState extends State<FynoNotificationIcon> {
   @override
   Widget build(BuildContext context) {
     widget.fynoInApp.stateUpdate = () {
-      setState(() {});
+      if (mounted) {
+        setState(() {});
+      }
     };
 
     return Stack(
       children: [
         IconButton(
           onPressed: () {
+            widget.fynoInApp.socket?.emit('updateLastSeen');
             Navigator.push(
               context,
               MaterialPageRoute(
                 builder: (context) => NotificationsPage(
                   widget.fynoInApp,
                   onUpdate,
+                  themeConfig: widget.themeConfig,
                 ),
               ),
             );
           },
           icon: Icon(
-            Icons.notifications,
+            widget.notificationIcon ?? Icons.notifications_outlined,
             color: widget.iconColor,
           ),
         ),
-        if (widget.fynoInApp.fynoInAppState.unreadCount > 0)
+        if (!widget.fynoInApp.fynoInAppState.isSeen)
           Positioned(
             right: 12,
             top: 10,
@@ -65,16 +89,23 @@ class FynoNotificationIconState extends State<FynoNotificationIcon> {
   }
 
   void onUpdate() {
-    setState(() {});
+    if (mounted) {
+      setState(() {});
+    }
   }
 }
 
 class NotificationsPage extends StatefulWidget {
   final FynoInApp fynoInApp;
   final Function onClick;
+  final ThemeConfig? themeConfig;
 
-  const NotificationsPage(this.fynoInApp, this.onClick, {Key? key})
-      : super(key: key);
+  const NotificationsPage(
+    this.fynoInApp,
+    this.onClick, {
+    this.themeConfig,
+    Key? key,
+  }) : super(key: key);
 
   @override
   NotificationsPageState createState() => NotificationsPageState();
@@ -167,21 +198,32 @@ class NotificationsPageState extends State<NotificationsPage> {
 
   AppBar buildAppBar(BuildContext context) {
     var unreadCount = widget.fynoInApp.fynoInAppState.unreadCount.toString();
+    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
 
     return AppBar(
       leading: BackButton(
         onPressed: () => {
           widget.onClick(),
           Navigator.pop(context),
-          setState(() {}),
+          if (mounted)
+            {
+              setState(() {}),
+            }
         },
       ),
-      iconTheme: IconThemeData(color: Colors.white),
-      backgroundColor: Theme.of(context).primaryColor,
+      iconTheme: IconThemeData(
+          color: isDarkMode
+              ? widget.themeConfig?.darkText ?? Colors.white
+              : widget.themeConfig?.lightText ?? Colors.black),
+      backgroundColor: isDarkMode
+          ? widget.themeConfig?.darkBackground ?? Colors.black
+          : widget.themeConfig?.lightBackground ?? Colors.white,
       title: Text(
         'Notifications',
         style: TextStyle(
-          color: Colors.white,
+          color: isDarkMode
+              ? widget.themeConfig?.darkText ?? Colors.white
+              : widget.themeConfig?.lightText ?? Colors.black,
         ),
       ),
       actions: [
@@ -213,8 +255,12 @@ class NotificationsPageState extends State<NotificationsPage> {
           ),
       ],
       bottom: TabBar(
-        labelColor: Colors.white,
-        unselectedLabelColor: Colors.white,
+        labelColor: isDarkMode
+            ? widget.themeConfig?.darkText ?? Colors.white
+            : widget.themeConfig?.lightText ?? Colors.black,
+        unselectedLabelColor: isDarkMode
+            ? widget.themeConfig?.darkText ?? Colors.white
+            : widget.themeConfig?.lightText ?? Colors.black,
         tabs: [
           Tab(text: 'All'),
           Tab(text: 'Unread ($unreadCount)'),
@@ -227,12 +273,16 @@ class NotificationsPageState extends State<NotificationsPage> {
     switch (action) {
       case 'markAllAsRead':
         widget.fynoInApp.markAllAsRead().then((_) {
-          setState(() {});
+          if (mounted) {
+            setState(() {});
+          }
         });
         break;
       case 'deleteAll':
         widget.fynoInApp.deleteAllMessages().then((_) {
-          setState(() {});
+          if (mounted) {
+            setState(() {});
+          }
         });
         break;
     }
@@ -286,6 +336,7 @@ class NotificationsPageState extends State<NotificationsPage> {
         onClick,
         tabName,
         scrollController,
+        themeConfig: widget.themeConfig,
       ),
     );
   }
@@ -297,6 +348,7 @@ class NotificationsTab extends StatelessWidget {
   final Function() onClick;
   final String tabName;
   final ScrollController? scrollController;
+  final ThemeConfig? themeConfig;
 
   const NotificationsTab(
     this.fynoInApp,
@@ -304,6 +356,7 @@ class NotificationsTab extends StatelessWidget {
     this.onClick,
     this.tabName,
     this.scrollController, {
+    this.themeConfig,
     Key? key,
   }) : super(key: key);
 
@@ -350,6 +403,7 @@ class NotificationsTab extends StatelessWidget {
       fynoInApp,
       message,
       onClick,
+      themeConfig: themeConfig,
     );
   }
 
@@ -372,11 +426,13 @@ class NotificationListTile extends StatefulWidget {
   final FynoInApp fynoInApp;
   final dynamic message;
   final Function() onClick;
+  final ThemeConfig? themeConfig;
 
   const NotificationListTile(
     this.fynoInApp,
     this.message,
     this.onClick, {
+    this.themeConfig,
     Key? key,
   }) : super(key: key);
 
@@ -409,10 +465,16 @@ class NotificationListTileState extends State<NotificationListTile> {
   }
 
   Widget buildListTile(BuildContext context, bool isUnread) {
+    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
+
     return ListTile(
       shape: Border(
         bottom: BorderSide(
-          color: Theme.of(context).secondaryHeaderColor.withOpacity(0.5),
+          color: isDarkMode
+              ? widget.themeConfig?.darkText?.withOpacity(0.5) ??
+                  Colors.white.withOpacity(0.5)
+              : widget.themeConfig?.lightText?.withOpacity(0.5) ??
+                  Colors.black.withOpacity(0.5),
         ),
       ),
       contentPadding: EdgeInsets.only(
@@ -420,14 +482,25 @@ class NotificationListTileState extends State<NotificationListTile> {
         left: 20,
         right: 20,
       ),
-      tileColor:
-          isUnread ? Theme.of(context).primaryColor.withOpacity(0.2) : null,
+      tileColor: isUnread
+          ? isDarkMode
+              ? widget.themeConfig?.darkBackground?.withOpacity(0.6) ??
+                  Colors.white.withOpacity(0.1)
+              : widget.themeConfig?.lightBackground?.withOpacity(0.6) ??
+                  Colors.black.withOpacity(0.1)
+          : Theme.of(context).brightness == Brightness.dark
+              ? widget.themeConfig?.darkBackground?.withOpacity(0.2) ??
+                  Colors.black
+              : widget.themeConfig?.lightBackground?.withOpacity(0.2) ??
+                  Colors.white,
       title: buildTitle(context),
     );
   }
 
   Widget buildTitle(BuildContext context) {
     var notificationContent = widget.message['notification_content'];
+    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
+
     return Stack(
       children: [
         Row(
@@ -443,6 +516,9 @@ class NotificationListTileState extends State<NotificationListTile> {
                     Text(
                       notificationContent['title'].toString(),
                       style: TextStyle(
+                        color: isDarkMode
+                            ? widget.themeConfig?.darkText ?? Colors.white
+                            : widget.themeConfig?.lightText ?? Colors.black,
                         fontWeight: FontWeight.bold,
                         fontSize: 15,
                       ),
@@ -451,6 +527,9 @@ class NotificationListTileState extends State<NotificationListTile> {
                     Text(
                       notificationContent['body'].toString(),
                       style: TextStyle(
+                        color: isDarkMode
+                            ? widget.themeConfig?.darkText ?? Colors.white
+                            : widget.themeConfig?.lightText ?? Colors.black,
                         height: 1.5,
                         fontWeight: FontWeight.normal,
                         fontSize: 12,
@@ -463,11 +542,11 @@ class NotificationListTileState extends State<NotificationListTile> {
                       children: [
                         Text(
                           GetTimeAgo.parse(
-                            DateTime.fromMillisecondsSinceEpoch(
-                                widget.message['status'][0]['timestamp']),
-                          ),
+                              DateTime.parse(widget.message['createdAt'])),
                           style: TextStyle(
-                            color: Colors.grey,
+                            color: isDarkMode
+                                ? widget.themeConfig?.darkText ?? Colors.white
+                                : widget.themeConfig?.lightText ?? Colors.black,
                             fontSize: 10,
                             fontWeight: !widget.message['isRead']
                                 ? FontWeight.bold
@@ -540,6 +619,7 @@ class NotificationListTileState extends State<NotificationListTile> {
 
   Widget buildButton(dynamic button) {
     bool isPrimary = button['primary'].toLowerCase() == "true";
+    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
 
     return Container(
       height: 25,
@@ -550,12 +630,20 @@ class NotificationListTileState extends State<NotificationListTile> {
         },
         style: ButtonStyle(
           backgroundColor: isPrimary
-              ? MaterialStateProperty.all(Theme.of(context).primaryColor)
+              ? MaterialStatePropertyAll(widget.themeConfig?.primary ??
+                  (isDarkMode ? Colors.white : Colors.black))
               : null,
           shape: MaterialStateProperty.all(
             RoundedRectangleBorder(
               borderRadius: BorderRadius.circular(4.0),
-              side: BorderSide(color: Theme.of(context).primaryColor),
+              side: BorderSide(
+                color: isPrimary
+                    ? widget.themeConfig?.primary ??
+                        (isDarkMode ? Colors.white : Colors.black)
+                    : isDarkMode
+                        ? widget.themeConfig?.darkText ?? Colors.white
+                        : widget.themeConfig?.lightText ?? Colors.black,
+              ),
             ),
           ),
         ),
@@ -563,7 +651,13 @@ class NotificationListTileState extends State<NotificationListTile> {
           button['label'].toString().toUpperCase(),
           style: TextStyle(
             fontSize: 8,
-            color: isPrimary ? Colors.white : Theme.of(context).primaryColor,
+            color: isPrimary
+                ? isDarkMode
+                    ? widget.themeConfig?.lightText ?? Colors.black
+                    : widget.themeConfig?.darkText ?? Colors.white
+                : isDarkMode
+                    ? widget.themeConfig?.darkText ?? Colors.white
+                    : widget.themeConfig?.lightText ?? Colors.black,
           ),
         ),
       ),
@@ -607,6 +701,8 @@ class NotificationListTileState extends State<NotificationListTile> {
     String videoUrl =
         widget.message['notification_content']['attachments']['attachment'];
 
+    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
+
     return SizedBox(
       height: 55,
       width: 55,
@@ -618,12 +714,18 @@ class NotificationListTileState extends State<NotificationListTile> {
           borderRadius: BorderRadius.circular(6.0),
           child: Container(
             padding: EdgeInsets.all(12),
-            color: Theme.of(context).primaryColor.withOpacity(0.15),
+            color: isDarkMode
+                ? widget.themeConfig?.darkBackground?.withOpacity(0.6) ??
+                    Colors.white.withOpacity(0.2)
+                : widget.themeConfig?.lightBackground?.withOpacity(0.6) ??
+                    Colors.black.withOpacity(0.2),
             child: Column(
               children: [
                 Icon(
                   Icons.play_circle,
-                  color: Theme.of(context).primaryColor,
+                  color: isDarkMode
+                      ? widget.themeConfig?.darkText ?? Colors.white
+                      : widget.themeConfig?.lightText ?? Colors.black,
                   size: 30,
                 ),
               ],
@@ -637,6 +739,7 @@ class NotificationListTileState extends State<NotificationListTile> {
   Widget buildDocument() {
     String documentUrl =
         widget.message['notification_content']['attachments']['attachment'];
+    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
 
     String fileExtension = path.extension(path.basename(documentUrl));
 
@@ -655,18 +758,26 @@ class NotificationListTileState extends State<NotificationListTile> {
           borderRadius: BorderRadius.circular(6.0),
           child: Container(
             padding: EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-            color: Theme.of(context).primaryColor.withOpacity(0.15),
+            color: isDarkMode
+                ? widget.themeConfig?.darkBackground?.withOpacity(0.6) ??
+                    Colors.white.withOpacity(0.2)
+                : widget.themeConfig?.lightBackground?.withOpacity(0.6) ??
+                    Colors.black.withOpacity(0.2),
             child: Column(
               children: [
                 Icon(
                   Icons.file_copy,
-                  color: Theme.of(context).primaryColor,
+                  color: isDarkMode
+                      ? widget.themeConfig?.darkText ?? Colors.white
+                      : widget.themeConfig?.lightText ?? Colors.black,
                   size: 20,
                 ),
                 Text(
                   fileType,
                   style: TextStyle(
-                    color: Theme.of(context).primaryColor,
+                    color: isDarkMode
+                        ? widget.themeConfig?.darkText ?? Colors.white
+                        : widget.themeConfig?.lightText ?? Colors.black,
                     fontSize: 10,
                   ),
                 ),
@@ -699,7 +810,9 @@ class NotificationListTileState extends State<NotificationListTile> {
                     width: 50,
                     height: 4.0,
                     decoration: BoxDecoration(
-                      color: Theme.of(context).secondaryHeaderColor,
+                      color: Theme.of(context).brightness == Brightness.dark
+                          ? widget.themeConfig?.darkText ?? Colors.white
+                          : widget.themeConfig?.lightText ?? Colors.black,
                       borderRadius: BorderRadius.circular(2.0),
                     ),
                   ),
@@ -749,7 +862,9 @@ class NotificationListTileState extends State<NotificationListTile> {
       child: IconButton(
         icon: Icon(
           Icons.more_horiz,
-          color: Theme.of(context).secondaryHeaderColor,
+          color: Theme.of(context).brightness == Brightness.dark
+              ? widget.themeConfig?.darkText ?? Colors.white
+              : widget.themeConfig?.lightText ?? Colors.black,
         ),
         onPressed: () {
           showBottomSheet(context, isUnread, message);
@@ -762,13 +877,17 @@ class NotificationListTileState extends State<NotificationListTile> {
     switch (value) {
       case 'markAsRead':
         widget.fynoInApp.markAsRead(message).then((_) {
-          setState(() {});
+          if (mounted) {
+            setState(() {});
+          }
           widget.onClick();
         });
         break;
       case 'delete':
         widget.fynoInApp.deleteMessage(message).then((_) {
-          setState(() {});
+          if (mounted) {
+            setState(() {});
+          }
           widget.onClick();
         });
         break;
@@ -778,7 +897,9 @@ class NotificationListTileState extends State<NotificationListTile> {
   void handleTap(bool isUnread) {
     if (isUnread) {
       widget.fynoInApp.markAsRead(widget.message).then((_) {
-        setState(() {});
+        if (mounted) {
+          setState(() {});
+        }
         widget.onClick();
       });
     }