+++ CHANGELOG.md
@@ -9,3 +9,31 @@
 ## 1.0.2
 
 - Added a parameter in callback function
+
+## 1.0.3
+
+- make onListUpdate function nullable
+
+## 1.0.4
+
+- added mounted check
+
+## 1.0.5
+
+- added mounted check for all setstate
+
+## 1.0.6
+
+- customisable notification icon and color
+
+## 1.0.7
+
+- added default logic to stateUpdate function to avoid error logs
+
+## 1.1.0
+
+- added reset function and seen check
+
+## 1.1.1
+
+- fixed UI rendering issue for null timestamp